+++ dialect/mysql/mysql_test.go
@@ -199,6 +199,24 @@ func (mt *mysqlTest) TestQuery() {
 	assert.Len(t, entries, 0)
 }
 
+func (mt *mysqlTest) TestQuery_ValueExpressions() {
+	type wrappedEntry struct {
+		entry
+		BoolValue bool `db:"bool_value"`
+	}
+	expectedDate, err := time.Parse("2006-01-02 15:04:05", "2015-02-22 19:19:55")
+	mt.NoError(err)
+	ds := mt.db.From("entry").Select(goqu.Star(), goqu.V(true).As("bool_value")).Where(goqu.Ex{"int": 1})
+	var we wrappedEntry
+	found, err := ds.ScanStruct(&we)
+	mt.NoError(err)
+	mt.True(found)
+	mt.Equal(we, wrappedEntry{
+		entry{2, 1, 0.100000, "0.100000", expectedDate, false, []byte("0.100000")},
+		true,
+	})
+}
+
 func (mt *mysqlTest) TestCount() {
 	t := mt.T()
 	ds := mt.db.From("entry")