+++ dialect/sqlite3/sqlite3.go
@@ -16,6 +16,7 @@ func DialectOptions() *goqu.SQLDialectOptions {
 	opts.SupportsConflictUpdateWhere = false
 	opts.SupportsInsertIgnoreSyntax = true
 	opts.SupportsConflictTarget = false
+	opts.SupportsMultipleUpdateTables = false
 	opts.WrapCompoundsInParens = false
 
 	opts.PlaceHolderRune = '?'